requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
gspread==2.0.0
httplib2==0.11.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
line-bot-sdk==1.5.0
